refactor: reframe devframe positioning + rename Vite integration
Reframe the docs/skill to position devframe as a portable, framework-
and build-tool-agnostic asset rather than a single-tool surface that
"lacks" hub features. Vite DevTools is now described as one host
built on devframe, reached via the renamed `vite` adapter (was the
`kit` adapter); the docs no longer enumerate things devframe doesn't
provide and instead point at Vite DevTools or custom adapters as the
escape valve.

The previous `vite` SPA-mounting plugin is no longer an adapter — moved
to `devframe/helpers/vite` and the factory renamed `createVitePlugin`
→ `viteDevBridge`. Updated the `@devframes/nuxt` consumer, internal
JSDoc, error doc DF0033, tsnapi snapshots, `DevframeRuntime` type
literal, and rebranded the client-side `[Vite DevTools]` timeout error
to `[devframe]`. The `vite-devtools-auth` BroadcastChannel literal is
kept for cross-tab interop with the existing Vite DevTools auth page.
