+++ packages/devframe/src/adapters/_shared.ts
@@ -2,9 +2,9 @@ import type { DevframeDefinition, DevframeDeploymentKind } from '../types/devfra
 
 /**
  * Resolve the mount base path for a devframe's SPA. Hosted adapters
- * (`vite`, `kit`, `embedded`) default to `/__<id>/` so they don't
- * collide with the host app; standalone adapters (`cli`, `spa`,
- * `build`) default to `/` because they own the origin.
+ * (`vite`, `embedded`) default to `/__<id>/` so they don't collide
+ * with the host app; standalone adapters (`cli`, `spa`, `build`)
+ * default to `/` because they own the origin.
  *
  * The devframe author can override with `basePath` on the definition.
  */
